requirements.txt
# Streamlit Community Cloud 全栈依赖
# 包含：前端 (Streamlit) + 后端 (Google Gemini API)

streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0        # 用于读取 Excel 文件
altair>=5.0.0          # 图表（情感分布等）
google-genai>=1.0.0    # 核心分析服务
pydantic>=2.0.0        # 数据模型校验
python-dotenv>=1.0.0   # 环境变量支持（本地）
openai>=1.0.0          # OpenAI & DeepSeek 兼容接口支持
httpx>=0.25.0          # 异步 HTTP 请求支持
